stack/telegraf/telegraf.d/ucs-power-temp-util/requirements.txt
# Package dependencies for retrieval of UCS data.
#
# Copyright (c) 2022 Cisco and/or its affiliates. All rights reserved.
#
# Licensed under the Apache License, Version 2.0 (the "License");
# you may not use this file except in compliance with the License.
# You may obtain a copy of the License at
#
#     http://www.apache.org/licenses/LICENSE-2.0
#
# Unless required by applicable law or agreed to in writing, software
# distributed under the License is distributed on an "AS IS" BASIS,
# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
# See the License for the specific language governing permissions and
# limitations under the License.
#
# AUTHOR(s): Cristina Precup <cprecup@cisco.com>

pyyaml==6.0
imcsdk==0.9.12
ucsmsdk==0.9.13
